Spendwise_backend/requirements.txt
Django==5.0.*
djangorestframework==3.15.*
django-cors-headers==4.3.*
mysqlclient==2.2.* 

gunicorn==22.0.0 
python-decouple==3.8
dj-database-url==3.0.0